feat(tuners): Make BaseTuner the shared lifecycle with ONLINE knob view
Promote BaseTuner from a thin run() template into a complete shared tuner
lifecycle. The concrete setup() now drives the entire common bring-up: workload-type resolution, knob-space construction, worker-resource resolution, hardware-range resolution, workload/executor
bundle, metric config, environment factory, orchestrator wiring, instance
bring-up + verification, runtime knob pruning, session-environment assembly,
and drawing the initial design via the propose_initial_configs() hook.
Subclasses now only contribute strategy-specific proposal/eval/stop/best/
payload behavior; LHSDesignTuner is slimmed to those seams.

Add the ONLINE knob-view derivation as a mandatory shared-lifecycle step: in
ONLINE mode the full knob space is applied + restarted ONCE at bring-up, then
only runtime-safe knobs are refined via config RELOAD.
